Cleaning up redundancies

+++ SupportClass.cs
@@ -115,7 +115,7 @@ namespace ComponentAce.Compression.Libs.zlib
 				return -1;
                 
 			for(int i = start; i < start + bytesRead; i++)
-				target[i] = (byte)receiver[i];
+				target[i] = receiver[i];
                 
 			return bytesRead;
 		}
